feat: Implement CI using github actions
As a part of this, I changed how the BLAS dependencies are resolved.
This is so that we can accelerate the build for testing, and for static
builds we just build openblas from source.

In addition, I had to fix the issues with building on apple anyways,
which ended up defeating the point of this whole exercise. The solution
in this case is to use preprocessor if statements and feature test
macros to conditionally include code which does not use the new
features. The OSX code passes all tests, but I haven't rigorously
tested it.

+++ tests/src/csv.cpp
@@ -16,8 +16,8 @@ TEST(CSVTest, simple) {
   EXPECT_EQ(row.get<std::string>("a"), "1");
   EXPECT_EQ(row.get<std::string>("b"), "2");
 
-  EXPECT_EQ(row.get<size_t>("a"), 1);
-  EXPECT_EQ(row.get<size_t>("b"), 2);
+  EXPECT_EQ(row.get<uint64_t>("a"), 1);
+  EXPECT_EQ(row.get<uint64_t>("b"), 2);
 }
 
 TEST(CSVTest, errors) {
@@ -32,6 +32,6 @@ TEST(CSVTest, errors) {
   EXPECT_EQ(row.get<std::string>("a"), "c");
   EXPECT_EQ(row.get<std::string>("b"), "d");
 
-  EXPECT_ANY_THROW(row.get<size_t>("a"));
-  EXPECT_ANY_THROW(row.get<size_t>("b"));
+  EXPECT_ANY_THROW(row.get<uint64_t>("a"));
+  EXPECT_ANY_THROW(row.get<uint64_t>("b"));
 }
